Move Wear OS specific classes into wearos package
