+++ mobile/src/main/java/com/boswelja/devicemanager/watchmanager/connection/wearos/PreferenceChangeReceiver.kt
@@ -5,7 +5,7 @@
  * This file, and any part of the Wearable Extensions app/s cannot be copied and/or distributed
  * without permission from Jack Boswell (boswelja) <boswela@outlook.com>
  */
-package com.boswelja.devicemanager.receiver
+package com.boswelja.devicemanager.watchmanager.connection.wearos
 
 import android.content.Intent
 import com.boswelja.devicemanager.batterysync.BatterySyncWorker
