+++ mobile/src/main/java/com/boswelja/devicemanager/watchmanager/connection/wearos/WatchMessageReceiver.kt
@@ -5,7 +5,7 @@
  * This file, and any part of the Wearable Extensions app/s cannot be copied and/or distributed
  * without permission from Jack Boswell (boswelja) <boswela@outlook.com>
  */
-package com.boswelja.devicemanager.receiver
+package com.boswelja.devicemanager.watchmanager.connection.wearos
 
 import android.app.admin.DevicePolicyManager
 import android.content.Context
