+++ mobile/src/test/java/com/boswelja/devicemanager/watchmanager/connection/WearOSConnectionInterfaceTest.kt
@@ -11,6 +11,7 @@ import com.boswelja.devicemanager.common.connection.Messages.CLEAR_PREFERENCES
 import com.boswelja.devicemanager.common.connection.Messages.RESET_APP
 import com.boswelja.devicemanager.common.preference.SyncPreferences
 import com.boswelja.devicemanager.getOrAwaitValue
+import com.boswelja.devicemanager.watchmanager.connection.wearos.WearOSConnectionInterface
 import com.boswelja.devicemanager.watchmanager.item.Watch
 import com.google.android.gms.tasks.Tasks
 import com.google.android.gms.wearable.CapabilityClient
